% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19api.R
\name{CovidAPI}
\alias{CovidAPI}
\title{Calls Covid19 API}
\usage{
CovidAPI(path)
}
\arguments{
\item{path}{API path response}
}
\value{
Get json requested by path string
}
\description{
Calls Covid19 API
}
\examples{
CovidAPI("summary")
}
