% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19api.R
\name{GetAllData}
\alias{GetAllData}
\title{Get all data}
\usage{
GetAllData()
}
\value{
Data frame columns country, Province, latitude, longitude, date,
number of cases and status
}
\description{
Download all data for all countries.
}
\examples{
GetAllData()
}
