% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19api.R
\name{GetAvalaibleCountries}
\alias{GetAvalaibleCountries}
\title{Get Countries Available in database}
\usage{
GetAvalaibleCountries()
}
\value{
data.frame structure with country name, sluj name and provinces
}
\description{
This functions get all the countries in database, with the slug
name and provinces if avalaible.
}
\examples{
GetAvalaibleCountries()
}
