% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19api.R
\name{GetByCountry}
\alias{GetByCountry}
\title{Get cases by Country}
\usage{
GetByCountry(country.requested, status.requested, live = FALSE, total = FALSE)
}
\arguments{
\item{country.requested}{Country slug name choosed}

\item{status.requested}{Status requested, they could be confirmed, recovered or deaths}

\item{live}{If TRUE gets the lates cases from the country and status requested}

\item{total}{If TRUE returns all cases by type for a country from the first recorded case}
}
\value{
Data frame columns country, Province, latitude, longitude, date, number of cases and status
}
\description{
Get all by type cases for a country. Country must be the slug from
GetAvalaibleCountries() or GetCountrySummary(). Cases must be one of:
confirmed, recovered, deaths. When total parameter is TRUE the live
parameter is not necesary.
}
\examples{
GetByCountry(country.requested = 'mexico', status.requested = 'confirmed')
GetByCountry(country.requested = 'mexico', status.requested = 'confirmed', live = TRUE)
GetByCountry(country.requested = 'mexico', status.requested = 'confirmed', total = TRUE)
}
