% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19api.R
\name{GetCountrySummary}
\alias{GetCountrySummary}
\title{Get country summary}
\usage{
GetCountrySummary()
}
\value{
Data Frame with the
}
\description{
Obtain data summary of new and total cases per country updated daily and
parse extraction datetime and country data.
}
\examples{
GetCountrySummary()
}
