% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19api.R
\name{GetDBStats}
\alias{GetDBStats}
\title{Get Database stats}
\usage{
GetDBStats()
}
\value{
List with stats by each api request calls
}
\description{
Get use stats for database request
}
\examples{
GetDBStats()
}
