% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19api.R
\name{GetLiveCountry}
\alias{GetLiveCountry}
\title{Get Live cases for country}
\usage{
GetLiveCountry(
  country.requested,
  status.requested,
  after.date = NA,
  after.time = "00:00:00"
)
}
\arguments{
\item{country.requested}{Country slug name choosed}

\item{status.requested}{Status requested, they could be confirmed, recovered or deaths}

\item{after.date}{Initial date to download data}

\item{after.time}{Initial time to donload data, '00:00:00' is default}
}
\value{
Data frame columns country, Province, latitude, longitude, datetime, number of cases and status
}
\description{
Get all live cases by  type for a country. Country must be the slug from
GetAvalaibleCountries() or GetCountrySummary(). Cases must be one of:
confirmed, recovered, deaths.
It could be filter by date an time.
}
\examples{
GetLiveCountry(country.requested = 'mexico', status.requested = 'confirmed')
GetLiveCountry(
  country.requested = 'mexico',
  status.requested = 'confirmed',
  after.date = '2020-03-27'
)
GetLiveCountry(
  country.requested = 'mexico',
  status.requested = 'confirmed',
  after.date = '2020-03-27',
  after.time = '17:57:00'
)
}
