% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19api.R
\name{GetDayOne}
\alias{GetDayOne}
\title{Get DayOne cases}
\usage{
GetDayOne(
  country.requested = "mexico",
  status.requested = "confirmed",
  live = FALSE
)
}
\arguments{
\item{country.requested}{Country slug name choosed}

\item{status.requested}{Status requested, they could be confirmed, recovered or deaths}

\item{live}{If TRUE gets the lates cases from the country and status requested}
}
\value{
Data frame columns country, Province, latitude, longitude, date, number of cases and status
}
\description{
Get all cases by type and country from the first recorded case.
Country must be the slug from GetAvalaibleCountries() or GetCountrySummary(). Cases
must be one of: confirmed, recovered, deaths.
}
\examples{
GetDayOne(country.requested = 'mexico', status.requested = 'confirmed')
#' @examples GetDayOne(country.requested = 'mexico', status.requested = 'confirmed', live = TRUE)
}
